allow reusable flows with initial path for AuthPickerView
